// api/src/modules/schedule/constants/enums/errors.enum.ts
export enum ErrorMessage {
  SCHEDULE_ID_EMPTY_ERROR = 'Bạn vui lòng chọn [lịch công tác].',
  ID_NAN_ERROR = 'Giá trị [id] không hợp lệ.',
  NO_CONTENT = 'Không có dữ liệu hiển thị.',
  OPERATOR_SCHEDULE_ERROR = 'Lưu thông tin [lịch công tác] thất bại.',
  FILE_NOT_FOUND_ERROR = 'Không tìm thấy file.',
  SCHEDULE_NOT_FOUND_ERROR = '[Lịch công tác] không tồn tại (id: %s)',
}